Fully test getPageIdFromPersistentId
Follow up to https://github.com/ProfessionalWiki/PersistentPageIdentifiers/pull/64

@alistair3149 Writing good tests after the fact is more difficult
than if you follow the red-green-refactor cycle. You never had the
red step for this test, because I could remove the WHERE clause
from the query, and it would still pass. To know a test actually
does what you want, you need to see it fail first. It's easy to
miss testing parts of the production code or to fail to test it
altogether due to a mistake in the test.